+++ app/NX/Flash/index.tsx
@@ -1,8 +1,8 @@
 import Flash from './Flash';
 import MovieClip from './MovieClips/MovieClip';
 import FlashDemo from './FlashDemo';
-import { setFlash } from './ActionScript/setFlash';
-import { useFlash } from './hooks/useFlash';
+import { setFlash } from './lib/actions/setFlash';
+import { useFlash } from './lib/hooks/useFlash';
 import Macromedia from './MovieClips/Icons/Macromedia';
 import LightningBolt from './MovieClips/Lightning/LightningBolt';
 import AnimateFlashLogo from './ActionScript/AnimateFlashLogo';
