+++ app/NX/Flash/lib/Flash.module.css
@@ -1,0 +1,21 @@
+/*
+    NOTE: For the Flash stage to fill the viewport, ensure your global CSS includes:
+        html, body {
+            height: 100%;
+            margin: 0;
+            padding: 0;
+        }
+    This cannot be set in a CSS module due to Next.js/webpack restrictions.
+*/
+
+/* Flash.module.css - ensures the Flash stage fills the viewport */
+.FlashStage {
+    width: 100vw;
+    height: 100vh;
+    min-height: 100vh;
+    min-width: 100vw;
+    position: relative;
+    overflow: hidden;
+    z-index: 0;
+    overscroll-behavior: none;
+}
