+++ app/NX/Flash/lib/hooks/useFlash.tsx
