+++ app/NX/Flash/lib/hooks/useIsMobile.tsx
@@ -1,4 +1,3 @@
-// ella/next/src/gl-core/cartridges/System/hooks/useIsMobile.tsx
 /*
     Hook returning whether we're on mobile
     determined by whether MUI size is < "md"
